% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{getmode}
\alias{getmode}
\title{Mode}
\usage{
getmode(v)
}
\arguments{
\item{v}{this is the vector of objects needing a mode}
}
\value{
most often data value of a given vector v.
}
\description{
This function will get the mode of a dataset.
}
\details{
R has no built in function for the mode of a vector so this function calculates the most often value in a vector.
}
\examples{
\dontrun{
mode <- getmode (c(4, 6, 6, 8))
mode holds the value of 6.
}
}
